projects/lorenz_attractor_lab/requirements.txt
numpy
scipy
pandas
matplotlib
pyarrow
pyyaml
